+++ src/raagas/random.rs
@@ -5,7 +5,7 @@ use crate::raagas::elements::{Pitch};
 use rand::seq::SliceRandom;
 
 
-pub fn randomiser(n: u32) -> Vec<Swar> {
+pub fn randomiser(n: usize) -> Vec<Swar> {
     let swars: Vec<String> = SWARS.keys().map(|x| x.to_string()).collect();
     let mut rnd = rand::thread_rng();
     let beats: Vec<Swar> = swars.choose_multiple(&mut rnd, n as usize)
